kotlin-racing-game : [refactor] 네이밍 수정
