+++ kotlin-racing-game/src/main/kotlin/racinggame/model/PositionRecorder.kt
@@ -1,18 +1,18 @@
 package racinggame.model
 
-data class PositionRecord(val positions: List<Position>) {
+class PositionRecorder(val positions: List<Position>) {
 
     fun getWinners(): List<String> {
         if (positions.isEmpty()) return emptyList()
         val maxPosition = findMaxPosition()
-        return findWinners(maxPosition)
+        return findNamesOfWinner(maxPosition)
     }
 
     private fun findMaxPosition(): Position =
         positions
             .maxBy { it.position }!!
 
-    private fun findWinners(maxPosition: Position): List<String> =
+    private fun findNamesOfWinner(maxPosition: Position): List<String> =
         positions
             .filter(maxPosition::isSamePosition)
             .map { it.driver }
